pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "forge-prime"
version = "1.0.0"
description = "Autonomous engineering OS — runner, dashboard, wiki, quality gate"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.29.0",
  "httpx>=0.27.0",
  "structlog>=24.1.0",
  "pyyaml>=6.0.1",
]

[project.scripts]
forge = "bin.forge:main"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
